More sound object fixes, mostly
I think I also fixed an issue with a zone border polygon in “Revisions of the Past”, hopefully. Finally, I renamed “Houses of the Holy” to “Miranda, That Ghost Just Isn’t Holy Anymore,” which is more in keeping with the idea I have for the story if I ever get around to writing it. Next commit should involve bringing file names up to date with the current level names, but IDK if I’ll have time to finish that before work, so I’ll submit it later.