Sound fixes, rearranged Earth levels, fixed glitch with “Deadwing” wires *again* (hopefully)..
The previous order allowed the player an unintentional shortcut; if they took a key out of “Miranda, That Ghost Just Isn’t Holy Anymore” (formerly “Houses of the Holy”), they would be able to skip part of a future level at will. I’m not too big on that idea, so I’ve fixed it. “Miranda” is the final level in the entire scenario to feature such a key.

“Deadwing” keeps getting a glitch with its wires that makes it impossible to finish the level. I’ll have to overhaul the wires completely; I’ll probably end up using an Enforcer/Tick shooting gallery like the one at the end of “Burn Down the Mission”.