Minor changes, mostly
Altered item placement in Sketches and Anthems, removed a bit of 5D space in Pleiades’ Dust implemented due to limitations on the engine and reconnected the areas that had been altered in this way, textured the new segments of Cut Their Grain & fixed a sound object, and… I don’t remember for Revisions. Might’ve made one of the secrets moderately less esoteric.