#18 now magnetic; #18 staff faster; #34 random sound changes
It didn’t make sense to have an underground mine without breathable air that wasn’t a magnetic level. I also raised the firing speed of the staff because it didn’t make sense to have it slower than it was on the surrounding levels, and I felt it was a suitable concession given the difficulty of the level. I hope this works; I saved a backup of the physics model just in case it didn’t.

#34 had ship creaks playing so often they’d frequently interrupt each other, so I mostly lowered the frequency. I think I raised their pitch slightly and made a couple other small changes as well.