lots of changes to #10 and #43; worked #43 and #44 into main scenario
#43 now has a proper mission; as a result, I’ve worked it and #44 into the main chronology, necessitating changes to the terminal file of #24 as well. I have yet to renumber the levels; I may not until I feel the scenario is reasonably close to being finished. #43 still needs a lot of work; I’d say less than a quarter of the map is truly populated, and the weapons and ammo placement require a complete overhaul. However, I’ve made a lot of progress on it.

With #10 I’ve attempted to change the random sounds specifically for this level using MML, but Aleph One doesn’t seem to recognise the extra sound slots I added to the sounds file and, as a result, simply doesn’t play those sounds. (If you understand what’s going wrong, please contact me with an explanation.) On the other hand, the “underground explosions” sound was never actually supposed to appear on this map (it was supposed to be Tempus Irae’s Gregorian chant sound), so that’s still a mild improvement.