a lot of changes to #45; some minor changes to #00
With #45 I widened the staircase, tweaked the upper canopy’s geometry a bit, and altered monster and item placement. Not completely satisfied with it and will probably change it some more.

I’m planning to scrap #00 entirely and replace it with a better intro, but for now I made the automap less hideous. I hope. I’ll see if it works once I merge.