monster changes in #45; monster & item impassable polygons in #01
I didn’t feel the closing battle of #45 was challenging enough, so I beefed up the monsters. I’m still not 100% sure it’s sufficient, but I think it’s better.

RADIX reported a S’pht’Kr getting stuck in a wall somewhere in #01 so I changed a few polygons to monster & item impassable. IDK if this’ll fix it but I’ll hope so.