sound changes in #10 and #43. Shapes/sounds files updated
#10 has a new MML script that alters the random sound behaviour on the level. This reflects some sounds I’ve moved around. I’ll be uploading new versions of the shapes & sounds files to reflect this once I’ve finished moving them around; in the interim, I’m likely to add some new ambient sounds as well. For the time being, you’ll probably get some Marathon 1 door sounds playing randomly until I upload the new versions.

In #43 I was mostly removing some ambient sound objects. Those things add up fast; too many in a map this size and A1 crashes. It’s worked fine for awhile, but I’ll probably tinker some more with the sounds, and removing some sound objects reduces the chances I’ll get crashes in the future. (In the process, I added a few as well, but I lowered the number of objects by almost 20, so it should be fine.)