Updates to #43
I’ve moved the pattern buffer near the start in order to get rid of a polygon. Correspondingly, I’ve changed the terminal a bit. If the pagination is messed up I will revise it further. I’ve also removed several more sound objects and changed the outside wind sound to the “wind & thunder” sound from Tempus. I hope to be able to use the “night wind” sound in several parts of the map eventually, but it remains to be seen whether there are enough map objects for me to do this seamlessly without causing crashes.