reshuffled level order a bit further
Instead of using the order Another One in the Dark -> Dimensional Bleedthrough -> Slaughterhouse-Five -> Cel din urmă vis -> Kill Your Sons, I made it Another One in the Dark -> Slaughterhouse-Five -> Cel din urmă vis -> Dimensional Bleedthrough -> Kill Your Sons. This does cause two of the longest, hardest levels in the game to appear in a row, but it also means that people who currently have save files on Another One in the Dark will have to make fewer manual level shifts from the console to see every level in the game. Plus it’s fairly true to M2, where the three hardest levels in the game appear in sequence (Sorry Don’t Make It So; For Carnage, Apply Within; and Begging for Mercy Makes Me Angry! – two of which are remixed in the two levels in question).