completely overhauled scenario’s handling of MML
MML for the solo map is now stored in the text resources rather than there being an individual file for each level. The benefit here is that when I alter the MML code for one specific task (e.g., using the Jjaro scenery in the sewage collection), I only have to alter it for one file. The disadvantage is that it’s not clear where everything is. Text resource 0 is currently an index of the scenario’s MML.

I was hoping to be able to do the same thing for Lua, but I was unable to figure it out at this late hour.

I also was testing out music for the level “Take the Veil Cerpin Taxt”. It turns out that my remastered version of the song is so quiet that you can barely hear it in-game, even after turning the volume all the way up. Another casualty of the loudness war, it seems. Of course, the original song had itself been loudness war’d, too (thanks, Rick Rubin!)